+++ Kickstarter-iOS/Features/SortFilter/HeaderView/SearchFiltersHeaderView.swift
@@ -44,6 +44,7 @@ struct SearchFiltersHeaderView: View {
         trailing: Styles.grid(1)
       ))
     }
+    .scrollIndicators(.never)
   }
 }
